backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "crewai[tools]>=0.186.1",
    "crewai-tools[mcp]>=0.1.0",
    "fastmcp>=2.12.3",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-socketio>=5.4.1",
    "langchain>=0.3.27",
    "libsql>=0.1.11",
    "mcp>=1.14.0",
    "netcdf4>=1.7.2",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.43",
    "tenacity>=9.1.2",
    "xarray>=2025.9.0",
    "pydantic>=2.11.9",
    "python-socketio>=5.13.0",
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "tqdm>=4.67.1",
    "requests>=2.32.5",
    "chromadb>=0.5.23",
    "werkzeug>=3.1.1",
    "sentence-transformers>=5.1.1",
]
